+++ CHANGELOG.md
@@ -5,11 +5,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 
 ## [Unreleased]
 
+## [2023.11.0]
+
 ### Fixed
 - In config, the channel range field is now used by the daemon.
 
 ### Changed
 - In config, channel range fields is now a list of floats (`[min, max]`).
+- Use `importlib` instead of `imp`, which was removed in python 3.12.
 
 ## [2022.11.0]
 
@@ -92,7 +95,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/).
 ### Added
 - initial release
 
-[Unreleased]: https://github.com/yaq-project/yaqd-ni/-/compare/v2022.11.0...main
+[Unreleased]: https://github.com/yaq-project/yaqd-ni/-/compare/v2023.11.0...main
+[2023.11.0]: https://github.com/yaq-project/yaqd-ni/-/compare/v2022.11.0....v2023.11.0
 [2022.11.0]: https://github.com/yaq-project/yaqd-ni/-/compare/v2022.7.1....v2022.11.0
 [2022.7.1]: https://github.com/yaq-project/yaqd-ni/-/compare/v2022.7.0....v2022.7.1
 [2022.7.0]: https://github.com/yaq-project/yaqd-ni/-/compare/v2022.2.0....v2022.7.0